move test classes out of 'co.elastic.apm' package
