+++ apm-agent-plugins/apm-jms-plugin/apm-jms-javax/src/test/java/testapp/TestMessageConsumer.java
@@ -16,12 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.agent.jms.jakarta.test;
+package testapp;
 
-import jakarta.jms.JMSException;
-import jakarta.jms.Message;
-import jakarta.jms.MessageConsumer;
-import jakarta.jms.MessageListener;
+import javax.jms.JMSException;
+import javax.jms.Message;
+import javax.jms.MessageConsumer;
+import javax.jms.MessageListener;
 
 // no-op implementation that is required only for testing registration of lambda message listeners
 public class TestMessageConsumer implements MessageConsumer {
